Fix remaining issues from PR review: update import paths, fix circular dependencies, improve state management
