+++ src/cli/benchmark/main.py
@@ -26,14 +26,14 @@ from src.cli.common.config import config
 from src.cli.main import console
 
 # Import artifact_guard utilities
-from src.artifact_guard import (
+from src.core.artifact_guard import (
     get_canonical_artifact_path,
     validate_artifact_path,
     PathGuard
 )
 
 # Import local modules
-from src.fastvlm_analyzer import FastVLMAnalyzer
+from src.models.fastvlm.analyzer import FastVLMAnalyzer
 
 # Check if PIL is available
 try:
@@ -159,37 +159,20 @@ def download_test_images(output_dir=None):
     
     return downloaded_paths
 
+# Import shared utilities
+from src.cli.benchmark.utils import find_test_images as find_test_images_util
+
 def find_test_images():
     """
     Find test images in canonical artifact paths and test_data directory.
-    
-    Attempts multiple strategies to find suitable test images:
-    1. Checks canonical benchmark test_images directory
-    2. Checks test_data/images directory
-    3. Downloads sample images if none found
-    
-    Returns:
-        List of Path objects pointing to test images
+    This is a wrapper around the utility function that adds console output.
     """
-    # Look in canonical benchmark path
-    benchmark_path = get_canonical_artifact_path("benchmark", "test_images")
-    image_list = []
+    # Get images from the utility function
+    image_list = find_test_images_util()
     
-    # Check if directory exists and contains images
-    if os.path.exists(benchmark_path):
-        for ext in [".jpg", ".jpeg", ".png", ".gif", ".tiff", ".bmp"]:
-            image_list.extend(list(Path(benchmark_path).glob(f"*{ext}")))
-    
-    # If no images found in benchmark path, check test_data directory
+    # Add console output for the CLI version
     if not image_list:
-        # Get project root
-        project_root = Path(__file__).resolve().parents[3]
-        test_path = project_root / "test_data" / "images"
-        
-        if test_path.exists():
-            console.print(f"[yellow]No images found in benchmark directory. Checking {test_path}...[/yellow]")
-            for ext in [".jpg", ".jpeg", ".png", ".gif", ".tiff", ".bmp"]:
-                image_list.extend(list(test_path.glob(f"*{ext}")))
+        console.print(f"[yellow]No images found in benchmark directories.[/yellow]")
     
     # If still no images found, download sample images
     if not image_list:
@@ -199,29 +182,8 @@ def find_test_images():
     
     return image_list
 
-def format_size(path):
-    """Format file size nicely"""
-    size_bytes = os.path.getsize(path)
-    for unit in ['B', 'KB', 'MB', 'GB']:
-        if size_bytes < 1024 or unit == 'GB':
-            return f"{size_bytes:.2f} {unit}"
-        size_bytes /= 1024
-
-def get_image_info(image_path):
-    """Get image dimensions and size"""
-    if not PIL_AVAILABLE:
-        return {"size": format_size(image_path), "dimensions": "Unknown (PIL not available)"}
-    
-    try:
-        with Image.open(image_path) as img:
-            width, height = img.size
-            return {
-                "size": format_size(image_path),
-                "dimensions": f"{width}x{height}",
-                "format": img.format
-            }
-    except Exception as e:
-        return {"size": format_size(image_path), "dimensions": f"Error: {str(e)}"}
+# Import additional shared utilities
+from src.cli.benchmark.utils import get_image_info, format_size
 
 def run_benchmark(analyzer, images, output_file=None):
     """
